cmake: version libceph-common to appease dpkg-shlibdeps
when packaging debian packages, dpkg-shlibdeps analyzes the linked
shared libraries of the binaries in given package by looking at their
names. but if the name does not include any useful versioning
information, it complains. we have no intention of versioning
libceph-common, as it is merely an internal shared library used by ceph
packages only. but there is no simple way to disable dpkg-shlibdeps'
warning of

  dpkg-shlibdeps: warning: can't extract name and version from library
  name 'libceph-common.so'

other than skipping the dpkg-shlibdeps for the whole package. so let's
just version it anyway.

Signed-off-by: Kefu Chai <kchai@redhat.com>